+++ packages/chameleon-mixins/wx-alipay-common-mixins.js
@@ -2,7 +2,7 @@
 
 const common = require('./common.js');
 const wxStyleHandle = require('chameleon-css-loader/proxy/proxyMiniapp.js')
-
+const utils = require('./utils.js')
 const deepClone = function(obj) {
   if (obj.toString().slice(8, -1) !== "Object") {
     return obj;
@@ -23,22 +23,24 @@ _.mixins = {
     // 支持事件传参
     [_.inlineStatementEventProxy](e) {
       let { dataset } = e.currentTarget;
-      let originFuncName = dataset && dataset[`event${e.type}`];
-
-      let argsStr = dataset && dataset.args;
-      let argsArr = argsStr.split(',').reduce((result, item, index) => {
-        let arg = dataset[`arg${index}`];
-        if (arg === "$event") {
-          let newEvent = getNewEvent(e);
-          result.push(newEvent);
-        } else {
-          // 这里的值微信已经计算好了；到dateset的时候已经是计算的结果 比如msg = 'sss' data-arg1="{{msg + 1}}"
-          // dataset[arg1] = 'sss1'
-          result.push(dataset[`arg${index}`])
-        }
-        return result;
-
-      }, []);
+      // 支付宝的e.type = 'touchStart',需要改为小写，否则找不到函数
+      e.type = utils.handleEventType(e.type);
+      let eventKey = e.type.toLowerCase();
+      let originFuncName = dataset && dataset[`event${eventKey}`] && dataset[`event${eventKey}`][0];
+      let inlineArgs = dataset && dataset[`event${eventKey}`] && dataset[`event${eventKey}`].slice(1);
+      let argsArr = [];
+      // 由于百度对于 data-arg="" 在dataset.arg = true 值和微信端不一样所以需要重新处理下这部分逻辑
+      if (inlineArgs) {
+        argsArr = inlineArgs.reduce((result, arg, index) => {
+          if (arg === "$event") {
+            let newEvent = getNewEvent(e);
+            result.push(newEvent);
+          } else {
+            result.push(arg)
+          }
+          return result;
+        }, []);
+      }
       if (originFuncName && this[originFuncName] && _.isType(this[originFuncName], 'Function')) {
         this[originFuncName](...argsArr)
       } else {
@@ -55,7 +57,10 @@ _.mixins = {
     },
     [_.eventProxyName](e) {
       let { dataset } = e.currentTarget;
-      let originFuncName = dataset && dataset[`event${e.type}`]
+      // 支付宝的e.type = 'touchStart',需要改为小写，否则找不到函数
+      e.type = utils.handleEventType(e.type);
+      let eventKey = e.type.toLowerCase();
+      let originFuncName = dataset && dataset[`event${eventKey}`] && dataset[`event${eventKey}`][0];
       if (originFuncName && this[originFuncName] && _.isType(this[originFuncName], 'Function')) {
         let newEvent = getNewEvent(e);
         this[originFuncName](newEvent)