+++ packages/chameleon-template-parse/test/index.js
@@ -1,6 +1,7 @@
 const compileTemplate = require('../src/index.js');
-const source = `<view c-bind:click="handleClick(1,2,3)" style="width:400cpx;height:200cpx;background-color:red">change-component
-<view c-bind:click="handleBubble" style="width:200cpx;height:50cpx;background-color:green"> 触发冒泡</view>
+const source = `<view class="demo-com" c-bind:click="handleClick" c-bind:tap="handleTap">
+<thirdComp2 c-bind:click="handleClick" c-bind:tap="handleTap">index-handleTouchStart</thirdComp2>
+<cube-button c-bind:click="handleClick" c-bind:tap="handleTap">index-handleTouchStart</cube-button>
 </view>`
 // <view><text :class="{{true? 'bg-green':''}}" >fafafa</text></view>
 // <view><text :class="true? 'bg-green':''" >fafafa</text></view>
@@ -9,6 +10,7 @@ const source = `<view c-bind:click="handleClick(1,2,3)" style="width:400cpx;heig
 let options = {lang: 'cml',
   filePath: '/Users/didi/components.cml',
   buildInComponents: {button: "cml-buildin-button"},
+  isInjectBaseStyle: false,
   cmss: {
     rem: true,
     scale: 0.5,
@@ -26,7 +28,7 @@ let options = {lang: 'cml',
     tagName: 'cube-button',
     refUrl: '/path/to/ref1',
     filePath: 'path/to/real1',
-    isNative: false
+    isNative: true
   }, {
     tagName: 'thirdComp2',
     refUrl: '/path/to/ref2',
@@ -35,12 +37,13 @@ let options = {lang: 'cml',
   }]
 };
 console.log('before-compile', source);
-// let result_web = compileTemplate(source, 'web', options);
+let result_web = compileTemplate(source, 'web', options);
 // let result_weex = compileTemplate(source, 'weex', options);
-let result_wx = compileTemplate(source, 'wx', options);
-// let result_baidu = compileTemplate(source, 'wx', options);
+// let result_wx = compileTemplate(source, 'wx', options);
+// let result_baidu = compileTemplate(source, 'baidu', options);
 // let result_alipay = compileTemplate(source, 'alipay', options);
-// console.log('result_web', result_web)
+console.log('result_web', result_web)
 // console.log('result_weex', result_weex)
-console.log('result_wx', result_wx)
+// console.log('result_wx', result_wx)
+// console.log('result_baidu', result_baidu)
 // console.log('result_alipay', result_alipay)