+++ packages/chameleon-tool/configs/getCommonConfig.js
@@ -25,8 +25,8 @@ module.exports = function (options) {
 
   let publicPath;
   let defaultPublichPathMap = {
-    'wx': '/',
-    'alipay': '/',
+    'wx': `http://${config.ip}:${webServerPort}/wx/`,
+    'alipay': `http://${config.ip}:${webServerPort}/alipay/`,
     'baidu': `http://${config.ip}:${webServerPort}/baidu/`, // baidu小程序的publicPath不能设置能/  所以在启动dev服务的时候 也将dist作为静态资源
     'web': `http://${config.ip}:${webServerPort}/`,
     'weex': `http://${config.ip}:${webServerPort}/weex/`
@@ -45,19 +45,19 @@ module.exports = function (options) {
       extensions: ['.cml', '.interface', '.vue', '.js'],
       alias: {
         '$CMLPROJECT': path.resolve(cml.root),
+        '/components': path.resolve(cml.projectRoot, 'src/components'),
         '$PROJECT': path.resolve(root),
-        '$ROUTER': path.resolve(root, 'node_modules/chameleon-runtime/.temp/router.js'),
         '$ROUTER_CONFIG': path.resolve(root, './src/router.config.json')
       },
       modules: [
         'node_modules',
-        path.join(cml.root, '/node_modules')
+        path.join(cml.root, '/node_modules'),
       ]
     },
     resolveLoader: {
       modules: [
-        'node_modules',
-        path.join(cml.root, '/node_modules')
+        path.join(cml.root, '/node_modules'),
+        'node_modules'
       ]
     },
     module: {
@@ -159,7 +159,6 @@ module.exports = function (options) {
   if (options.analysis) {
     commonConfig.plugins.push(new BundleAnalyzerPlugin())
   }
-
   let devApiPrefix = `http://${config.ip}:${webServerPort}`
   // 兼容旧版api
   let apiPrefix = options.apiPrefix || devApiPrefix;
@@ -213,7 +212,7 @@ module.exports = function (options) {
     commonConfig.plugins = commonConfig.plugins.concat([
       new OptimizeCSSPlugin({
         assetNameRegExp: /\.css$/,
-        cssProcessorOptions: { safe: true, discardComments: { removeAll: true } }
+        cssProcessorOptions: { safe: true, discardComments: { removeAll: true }, autoprefixer: false } 
       }),
       new UglifyJsPlugin({})
     ])
@@ -229,9 +228,9 @@ module.exports = function (options) {
     commonConfig.plugins.push(moduleIdMap[moduleIdType])
   }
 
-  let cmlPages = cml.config.get().cmlPages;
-  if (cmlPages && cmlPages.length > 0) {
-    cmlPages.forEach(npmName => {
+  let subProject = cml.config.get().subProject;
+  if (subProject && subProject.length > 0) {
+    subProject.forEach(npmName => {
       let packageJSON = JSON.parse(fs.readFileSync(path.resolve(cml.projectRoot, 'node_modules', npmName, 'package.json'),{encoding:'utf-8'}));
       let cmlConfig = packageJSON.cml || {};
       let definePlugin = cmlConfig.definePlugin;