+++ packages/chameleon-tool/configs/getMiniAppBuildConfig.js
@@ -1,15 +1,11 @@
 var OptimizeCSSPlugin = require('optimize-css-assets-webpack-plugin')
 var webpack = require('webpack');
-var UglifyJsPlugin = require('uglifyjs-webpack-plugin')
 var getMiniAppCommonConfig = require('./getMiniAppCommonConfig.js');
 var merge = require('webpack-merge')
 var getMiniAppExportConfig = require("./component_export/getMiniAppExportConfig");
 
 module.exports = function (options) {
   let {type, media} = options;
-  if (media === 'export') {
-    return getMiniAppExportConfig(options);
-  }
   var commonConfig = getMiniAppCommonConfig(options);
   var buildConfig = {
     plugins: [
@@ -32,10 +28,12 @@ module.exports = function (options) {
     buildConfig.plugins = [
       new OptimizeCSSPlugin({
         assetNameRegExp: targetObj.cssReg,
-        cssProcessorOptions: { safe: true, discardComments: { removeAll: true } }
-      }),
-      new UglifyJsPlugin({})
+        cssProcessorOptions: { safe: true, discardComments: { removeAll: true }, autoprefixer: false }
+      })
     ]
+  }
+  if (media === 'export') {
+    return getMiniAppExportConfig(merge(commonConfig, buildConfig), options);
   }
   return merge(commonConfig, buildConfig)
 }