+++ packages/chameleon-tool/configs/routerLoader.js
@@ -35,11 +35,11 @@ module.exports = function(content) {
         })
     }
 
-    // cmlPages 中的页面
-    let cmlPages = cml.config.get().cmlPages;
-    if (cmlPages && cmlPages.length > 0) {
-      cmlPages.forEach(function(npmName) {
-        let npmRouterConfig = cml.utils.readCmlPagesRouterConfig(cml.projectRoot, npmName);
+    // subProject 中的页面
+    let subProject = cml.config.get().subProject;
+    if (subProject && subProject.length > 0) {
+      subProject.forEach(function(npmName) {
+        let npmRouterConfig = cml.utils.readsubProjectRouterConfig(cml.projectRoot, npmName);
         npmRouterConfig.routes && npmRouterConfig.routes.forEach(item => {
           let cmlFilePath = path.join(cml.projectRoot, 'node_modules', npmName, 'src', item.path + '.cml');
           routerList += `